save network flags only in ntManager
